Move out of article namespaces (#582)
