+++ .gitattributes
@@ -5,5 +5,5 @@
 .gitattributes export-ignore
 .gitignore export-ignore
 .php_cs.dist export-ignore
-.phpstan.neon export-ignore
-.phpunit.xml.dist export-ignore
+phpstan.neon export-ignore
+phpunit.xml.dist export-ignore
