+++ Domain/Model/ArticleInterface.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\ArticleBundle\Article\Domain\Model;
+namespace Sulu\Bundle\ArticleBundle\Domain\Model;
 
 use Sulu\Bundle\ContentBundle\Content\Domain\Model\ContentRichEntityInterface;
 use Sulu\Component\Persistence\Model\AuditableInterface;
