+++ Resources/config/doctrine/Article/Article.orm.xml
@@ -3,7 +3,7 @@
                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
     <mapped-superclass
-        name="Sulu\Bundle\ArticleBundle\Article\Domain\Model\Article"
+        name="Sulu\Bundle\ArticleBundle\Domain\Model\Article"
         table="ar_articles"
         repository-class="Sulu\Component\Persistence\Repository\ORM\EntityRepository"
     >
@@ -11,7 +11,7 @@
             <generator strategy="NONE"/>
         </id>
 
-        <one-to-many field="dimensionContents" target-entity="Sulu\Bundle\ArticleBundle\Article\Domain\Model\ArticleDimensionContentInterface" mapped-by="article">
+        <one-to-many field="dimensionContents" target-entity="Sulu\Bundle\ArticleBundle\Domain\Model\ArticleDimensionContentInterface" mapped-by="article">
             <cascade>
                 <cascade-persist/>
             </cascade>
