+++ Resources/config/doctrine/Article/ArticleDimensionContent.orm.xml
@@ -2,7 +2,7 @@
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-    <entity name="Sulu\Bundle\ArticleBundle\Article\Domain\Model\ArticleDimensionContent"
+    <entity name="Sulu\Bundle\ArticleBundle\Domain\Model\ArticleDimensionContent"
             table="ar_articles_dimension_contents">
         <id name="id" type="integer" column="id">
             <generator strategy="AUTO"/>
@@ -10,7 +10,7 @@
 
         <field name="title" type="string" column="title" length="64" nullable="true"/>
 
-        <many-to-one field="article" target-entity="Sulu\Bundle\ArticleBundle\Article\Domain\Model\ArticleInterface" inversed-by="dimensionContents">
+        <many-to-one field="article" target-entity="Sulu\Bundle\ArticleBundle\Domain\Model\ArticleInterface" inversed-by="dimensionContents">
             <join-columns>
                 <join-column name="articleId" referenced-column-name="id" on-delete="CASCADE" nullable="false"/>
             </join-columns>
