+++ Tests/Application/config/config_experimental_storage.yml
@@ -1,3 +1,2 @@
 sulu_article:
-    article:
-        storage: 'experimental'
+    storage: 'experimental'
