+++ Tests/Unit/Domain/Model/ArticleTest.php
@@ -9,11 +9,11 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\ArticleBundle\Tests\Unit\Article\Domain\Model;
+namespace Sulu\Bundle\ArticleBundle\Tests\Unit\Domain\Model;
 
 use PHPUnit\Framework\TestCase;
-use Sulu\Bundle\ArticleBundle\Article\Domain\Model\Article;
-use Sulu\Bundle\ArticleBundle\Article\Domain\Model\ArticleInterface;
+use Sulu\Bundle\ArticleBundle\Domain\Model\Article;
+use Sulu\Bundle\ArticleBundle\Domain\Model\ArticleInterface;
 
 class ArticleTest extends TestCase
 {
