+++ Tests/Unit/Infrastructure/Sulu/Admin/ArticleAdminTest.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\ArticleBundle\Tests\Unit\Article\Infrastructure\Sulu\Admin;
+namespace Sulu\Bundle\ArticleBundle\Tests\Unit\Infrastructure\Sulu\Admin;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
@@ -19,8 +19,8 @@ use Sulu\Bundle\AdminBundle\Admin\View\ViewBuilderFactory;
 use Sulu\Bundle\AdminBundle\Admin\View\ViewBuilderFactoryInterface;
 use Sulu\Bundle\AdminBundle\Admin\View\ViewBuilderInterface;
 use Sulu\Bundle\AdminBundle\Admin\View\ViewCollection;
-use Sulu\Bundle\ArticleBundle\Article\Domain\Model\ArticleInterface;
-use Sulu\Bundle\ArticleBundle\Article\Infrastructure\Sulu\Admin\ArticleAdmin;
+use Sulu\Bundle\ArticleBundle\Domain\Model\ArticleInterface;
+use Sulu\Bundle\ArticleBundle\Infrastructure\Sulu\Admin\ArticleAdmin;
 use Sulu\Bundle\ContentBundle\Content\Infrastructure\Sulu\Admin\ContentViewBuilderFactoryInterface;
 use Sulu\Component\Localization\Manager\LocalizationManagerInterface;
 use Sulu\Component\Security\Authorization\PermissionTypes;
