+++ phpunit.xml.dist
@@ -3,8 +3,13 @@
          bootstrap="Tests/test-bootstrap.php"
 >
     <testsuites>
-        <testsuite name="Sulu Article Bundle">
-            <directory>./Tests</directory>
+        <!-- Run Unit Tests before Functional Tests for better Developer Experience -->
+        <testsuite name="Unit">
+            <directory>Tests/Unit</directory>
+        </testsuite>
+
+        <testsuite name="Functional">
+            <directory>Tests/Functional</directory>
         </testsuite>
     </testsuites>
 
@@ -20,8 +25,10 @@
     </filter>
 
     <php>
-        <env name="KERNEL_CLASS" value="Sulu\Bundle\ArticleBundle\Tests\Application\Kernel"/>
-        <env name="APP_ENV" value="test" force="true"/>
+        <ini name="error_reporting" value="-1" />
+        <server name="SHELL_VERBOSITY" value="-1" />
+        <server name="APP_ENV" value="test" force="true"/>
+        <server name="KERNEL_CLASS" value="Sulu\Bundle\ArticleBundle\Tests\Application\Kernel"/>
         <env name="SYMFONY_DEPRECATIONS_HELPER" value="weak"/>
     </php>
 </phpunit>
